+++ src/SFA.Apprenticeship.Api.AvService/Mappers/Version51/VacancyUploadRequestMapper.cs
@@ -14,7 +14,7 @@
 
     public class VacancyUploadRequestMapper : IVacancyUploadRequestMapper
     {
-        public ApprenticeshipVacancy ToVacancy(long vacancyReferenceNumber, VacancyUploadData vacancyUploadData, ProviderSiteEmployerLink providerSiteEmployerLink)
+        public ApprenticeshipVacancy ToVacancy(int vacancyReferenceNumber, VacancyUploadData vacancyUploadData, ProviderSiteEmployerLink providerSiteEmployerLink)
         {
             var ukprn = Convert.ToString(vacancyUploadData.ContractedProviderUkprn);
             var providerSiteErn = Convert.ToString(vacancyUploadData.VacancyOwnerEdsUrn);
