+++ src/SFA.Apprenticeships.Domain.Interfaces/Repositories/IReferenceNumberRepository.cs
@@ -2,6 +2,6 @@
 {
     public interface IReferenceNumberRepository
     {
-        long GetNextVacancyReferenceNumber();
+        int GetNextVacancyReferenceNumber();
     }
 }
