+++ src/SFA.Apprenticeships.Web.Manage.UnitTests/Mediators/Vacancy/VacancyMediatorTests.cs
@@ -21,7 +21,7 @@ namespace SFA.Apprenticeships.Web.Manage.UnitTests.Mediators.Vacancy
         [Test]
         public void ShouldUpdateTheStatusOfTheVacancyToLive()
         {
-            const long vacancyReferenceNumber = 1;
+            const int vacancyReferenceNumber = 1;
             var pendingQAVacancies = GetPendingVacancies(new[]
             {
                 vacancyReferenceNumber
@@ -40,7 +40,7 @@ namespace SFA.Apprenticeships.Web.Manage.UnitTests.Mediators.Vacancy
         public void ShouldUpdateTheStatusOfTheVacancyToDraft()
         {
             //Arrange
-            const long vacancyReferenceNumber = 1;
+            const int vacancyReferenceNumber = 1;
             var pendingQAVacancies = GetPendingVacancies(new[]
             {
                 vacancyReferenceNumber
@@ -60,9 +60,9 @@ namespace SFA.Apprenticeships.Web.Manage.UnitTests.Mediators.Vacancy
         [Test]
         public void ShouldReturnTheNextAvailableVacancyAfterApprovingOne()
         {
-            const long vacancyReferenceNumber = 1;
-            const long nextVacancyReferenceNumber = 2;
-            const long anotherVacancyReferenceNumber = 3;
+            const int vacancyReferenceNumber = 1;
+            const int nextVacancyReferenceNumber = 2;
+            const int anotherVacancyReferenceNumber = 3;
 
             var pendingQAVacancies = GetPendingVacancies(new[]
             {
@@ -84,7 +84,7 @@ namespace SFA.Apprenticeships.Web.Manage.UnitTests.Mediators.Vacancy
         [Test]
         public void ShouldReturnNoAvailableVacanciesIfThereArentAnyAvailableVacancies()
         {
-            const long vacancyReferenceNumber = 1;
+            const int vacancyReferenceNumber = 1;
 
             List<DashboardVacancySummaryViewModel> pendingQAVacancies = null;
 
@@ -98,7 +98,7 @@ namespace SFA.Apprenticeships.Web.Manage.UnitTests.Mediators.Vacancy
             result.AssertCode(VacancyMediatorCodes.ApproveVacancy.NoAvailableVacancies);
         }
 
-        private IEnumerable<DashboardVacancySummaryViewModel> GetPendingVacancies(IEnumerable<long> vacancyReferenceNumbers )
+        private IEnumerable<DashboardVacancySummaryViewModel> GetPendingVacancies(IEnumerable<int> vacancyReferenceNumbers )
         {
             return vacancyReferenceNumbers.Select(vacancyReferenceNumber => new DashboardVacancySummaryViewModel
             {
